Deleted the unused function to compute the quantile.
The previous method to comptue the quantile was a bit stupid. I've changed it with the classical pinball (or check) function. The loss is now the residual from the check function. Note that the weight of each observation is taken into account in the min_function. Obviously, some points that have small weight should have low impact on the minimisation.